--- openmc/deplete/integrator/__init__.py
@@ -1,17 +1,0 @@
-"""
-Integrator
-===========
-
-The integrator subcomponents.
-"""
-
-from .abc import Integrator
-from .cf4 import *
-from .cecm import *
-from .celi import *
-from .cram import *
-from .epc_rk4 import *
-from .leqi import *
-from .predictor import *
-from .si_celi import *
-from .si_leqi import *
