+++ tests/unit_tests/test_deplete_cram.py
@@ -6,7 +6,7 @@ Compares a few Mathematica matrix exponentials to CRAM16/CRAM48.
 from pytest import approx
 import numpy as np
 import scipy.sparse as sp
-from openmc.deplete.integrator import CRAM16, CRAM48
+from openmc.deplete.cram import CRAM16, CRAM48
 
 
 def test_CRAM16():
